@uppy/dashboard - factored out code related to focus into a separate file
